requirements.txt
Flask==2.3.3
opencv-python-headless==4.8.1.78
Pillow==10.0.1
scenedetect==0.6.2
gunicorn==21.2.0